fix: read env at runtime
